Renaming of function page
I renamed my f1 page to displayGameIntro to clarify. I then renamed my A4 code to clarify it's mine.

If you want a different A4 code, then use it, mine's just there for reference.
